requirements.txt
django>=6.0,<6.1
djangorestframework>=3.16,<3.17
djangorestframework-simplejwt>=5.5,<6.0
django-cors-headers>=4.9,<5.0
requests>=2.32,<3.0
